feat: code splitting
